preprod

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: recherra <recherra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/10 15:59:43 by recherra          #+#    #+#             */
-/*   Updated: 2023/12/28 17:07:31 by recherra         ###   ########.fr       */
+/*   Updated: 2025/02/17 15:46:00 by recherra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
 	int	i;
